Concretizer: Reimplement soft constraints
We no longer use the Optimizer() solver for soft constraints.
Rather, we introduce a fresh Boolean variable for each constraint and
initially assert all variables. If the resulting problem is unsat, we
randomly retract constraints one at a time until we get a satisfiable
problem.
This speeds up the concretization, and concretized variables still
appear to be selected uniformly at random.