Let apply_guard() overapproximate guarded state
When the guard evaluates to MAYBE on a state s that can no longer be
bisected (this can happen if the abstract domain loses enough info),
return s itself rather than raising an exception.

This means that apply_guard(g, s) is NO LONGER guaranteed to be
an abstract state s1 where g ALWAYS holds.
(It will depend on the domain)